[#229] fix: 파일 이름 오타 수정
